Update snack DB price to float

+++ db/migrate/20200410151025_create_snacks.rb
@@ -2,10 +2,9 @@ class CreateSnacks < ActiveRecord::Migration[5.1]
   def change
     create_table :snacks do |t|
       t.string :name
-      t.integer :price
+      t.float :price
 
       t.timestamps
     end
   end
 end
-
